+++ index.html
@@ -4,17 +4,17 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./src/static/reset.css" />
+    <link rel="stylesheet" href="./src/common/reset.css" />
     <title>우리동네 해결사</title>
     <style>
       @font-face {
         font-family: "NanumSquareRound";
-        src: url(./src/assets/fonts/NanumSquareRoundL.ttf);
+        src: url(./src/common/fonts/NanumSquareRoundL.ttf);
       }
 
       @font-face {
         font-family: "NanumSquareRoundB";
-        src: url(./src/assets/fonts/NanumSquareRoundB.ttf);
+        src: url(./src/common/fonts/NanumSquareRoundB.ttf);
       }
 
       :root {
