+++ src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter } from "react-router-dom"
 import Header from './Header/Header.tsx'
 import Footer from './Footer/Footer.tsx'
 import PageRouter from "./PageRouter.tsx"
-import { DeleteModal } from "./Pages/ViewPage.tsx"
+import { DeleteModal } from "./Pages/ViewPage/ViewPage.tsx"
 
 function App() {
