+++ src/Pages/MainPage/styled.ts
@@ -1,0 +1,13 @@
+import { styled } from "styled-components"
+
+export const MapBox = styled.div`
+  width: 100%;
+  height: 100%;
+  margin-bottom: 20%;
+  background-color: white;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`
+export const Box = styled.div``
