+++ src/Pages/ReviewErranderPage/ReviewErranderPage.tsx
@@ -1,0 +1,11 @@
+import { Container, Title} from './styled'
+
+
+export default function ReviewErranderPage() {
+  // 의뢰인 평가 ---> 채팅창에서 평가로 넘어올 수 있도록?
+  return (
+    <Container>
+      <Title>의뢰인 평가</Title>
+    </Container>
+  )
+}
