+++ src/Pages/ReviewErranderPage/styled.ts
@@ -1,0 +1,15 @@
+import { styled } from "styled-components"
+
+export const Container = styled.section`
+  height: 100vh;
+  margin: 60px 22px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`
+export const Title = styled.h3`
+  font-weight: 700;
+  font-size: 22px;
+  margin: 22px 0;
+`
