+++ src/Pages/ReviewSolverPage/ReviewSolverPage.tsx
@@ -1,0 +1,9 @@
+import { Container, Title} from './styled'
+
+export default function ReviewSolverPage() {
+  return (
+    <Container>
+      <Title>수행인 평가</Title>
+    </Container>
+  )
+}
