+++ src/Pages/SignpDetailPage/SingupDetailPage.tsx
@@ -1,8 +1,8 @@
 import axios from 'axios';
 import { useState, useCallback, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom';
-import styled from 'styled-components'
-import { REST_API_KEY, REDIRECT_URI } from '../kakaoLoginData';
+import { REST_API_KEY, REDIRECT_URI } from '../LoginPage/kakao/kakaoLoginData';
+import { Container, InputBox, Title, Box, EmailInputBox, EmailInput, EmailCheckBtn, Alertment, PhonenumberInput, MapBox, Map, UserLastBtn} from './styled'
 
 declare global {
   interface Window {
@@ -216,109 +216,6 @@ const SingupDetailPage: React.FC = () => {
     </Container>
   )
 }
-
-const Container = styled.section`
-  margin: 80px 22px;
-`
-const InputBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`
-const Title = styled.h1`
-  font-weight: 700;
-  font-size: 22px;
-  margin: 12px 0;
-`
-
-const Box = styled.div`
-  width: 85%;
-  display: flex;
-  align-items: baseline;
-  flex-direction: column;
-  .message {
-    &.success {
-      color: #0500ff;
-    }
-    &.error {
-      color: #ff2727;
-    }
-  }
-`
-
-const EmailInputBox = styled.div`
-  display: flex;
-  align-items: baseline;
-  width: 100%;
-`
-const EmailInput = styled.input`
-  width: 70%;
-  margin: 6px 0;
-  padding: 8px;
-  border-radius: 12px;
-  border: 0;
-
-  background-color: #F5F5F5;
-
-  outline: none;
-`
-
-const Alertment = styled.span`
-  font-size: 12px;
-  padding-left: 8px;
-`
-
-const EmailCheckBtn = styled.button`
-  margin-left: 8px;
-  padding: 10px;
-  border-radius: 8px;
-  border: 0;
-
-  background-color: #0089B5;
-  color: #fff;
-  
-  font-size: 12px;
-  font-weight: 500;
-`
-
-const PhonenumberInput = styled.input`
-  width: 95%;
-  margin: 6px 0;
-  padding: 8px;
-  border-radius: 12px;
-  border: 0;
-
-  background-color: #F5F5F5;
-
-  outline: none;
-`
-
-const MapBox = styled.div`
-  width: 100%;
-  height: 100%;
-  margin: 15px 0;
-  background-color: white;
-
-  display: flex;
-  justify-content: center;
-`
-const Map = styled.div``
-
-const UserLastBtn = styled.button`
-  width: 85%;
-  margin: 0 auto;
-  padding: 10px;
-  border-radius: 6px;
-  border: 0;
-
-  background-color: #0089B5;
-  color: #fff;
-
-  font-size: 16px;
-  font-weight: 700;
-
-  cursor: pointer;
-`
 
 export default SingupDetailPage;
