+++ src/Pages/SignupPage/SingupPage.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react"
-import styled from "styled-components"
 import { useNavigate } from 'react-router-dom';
+import { Container, InputBox, SingupTitle, Box, EmailInputBox, EmailInput, EmailCheckBtn, Alertment, PwInput, PwCheckInput, SingupBtn} from './styled'
 
 export default function SingupPage() {
   // 입력한 이메일, 비밀번호
@@ -122,103 +122,3 @@ export default function SingupPage() {
   )
 }
 
-const Container = styled.section`
-  margin: 80px 22px;
-`
-const InputBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`
-const SingupTitle = styled.h1`
-  font-weight: 700;
-  font-size: 22px;
-  margin: 12px 0;
-`
-const Box = styled.div`
-  width: 85%;
-  display: flex;
-  align-items: baseline;
-  flex-direction: column;
-  .message {
-    &.success {
-      color: #0500ff;
-    }
-    &.error {
-      color: #ff2727;
-    }
-  }
-`
-const EmailInputBox = styled.div`
-  display: flex;
-  align-items: baseline;
-  width: 100%;
-`
-const EmailInput = styled.input`
-  width: 70%;
-  margin: 6px 0;
-  padding: 8px;
-  border-radius: 12px;
-  border: 0;
-
-  background-color: #F5F5F5;
-
-  outline: none;
-`
-
-const Alertment = styled.span`
-  font-size: 12px;
-  padding-left: 8px;
-`
-
-const EmailCheckBtn = styled.button`
-  margin-left: 8px;
-  padding: 10px;
-  border-radius: 8px;
-  border: 0;
-
-  background-color: #0089B5;
-  color: #fff;
-  
-  font-size: 12px;
-  font-weight: 500;
-
-  cursor: pointer;
-`
-const PwInput = styled.input`
-  width: 95%;
-  margin: 6px 0;
-  padding: 8px;
-  border-radius: 12px;
-  border: 0;
-
-  background-color: #F5F5F5;
-
-  outline: none;
-`
-const PwCheckInput = styled.input`
-  width: 95%;
-  margin: 6px 0;
-  padding: 8px;
-  border-radius: 12px;
-  border: 0;
-
-  background-color: #F5F5F5;
-
-  outline: none;
-`
-const SingupBtn = styled.button`
-  width: 85%;
-  margin: 6px 0;
-  padding: 10px;
-  border-radius: 6px;
-  border: 0;
-
-  background-color: #0089B5;
-  color: #fff;
-
-  font-size: 16px;
-  font-weight: 700;
-
-  cursor: pointer;
-`
