+++ src/Pages/ViewPage/styled.ts
@@ -1,0 +1,156 @@
+import { styled } from "styled-components"
+
+export const Modal = styled.div`
+    width: 80%;
+    height: 30%;
+    background-color: white;
+    border-radius: 10px;
+`
+
+export const DoYouWantDelete = styled.div`
+    display: flex;
+    justify-content: center;
+    height: 50%;
+    align-items: center;
+`
+
+export const Choice = styled.div`
+    display: flex;
+    width: 100%;
+    height: 50%;
+`
+
+export const ChoiceButton = styled.button`
+    width: 50%;
+    font-size: 30px;
+    border: 1px solid black;
+    background-color: white;
+    cursor: pointer;
+`
+
+export const Container = styled.section`
+    height: 100vh;
+    width: 330px;
+    margin: 60px 22px;
+`;
+export const EditBox = styled.div`
+    width: 100%;
+    display: flex;
+    justify-content: flex-end;
+    margin-bottom: 20px;
+    `
+export const EditButton = styled.button`
+    border: none;
+    width: 15%;
+    cursor: pointer;
+    border-radius: 10%;
+`
+
+export const Image = styled.img`
+    width: 100%;
+    height: 330px;
+    margin-bottom: 20px;
+`
+
+export const ProfileBox = styled.div`
+    width: 100%;
+    height: 10%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 20px;
+`
+export const ProfileImage = styled.img`
+    width: 50px;
+    height: 50px;
+    border-radius: 100%;
+`
+export const ProfileSubBox = styled.div`
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    flex-grow: 1;
+    margin: 0 2px;
+`
+export const ProfileNickName = styled.h2`
+    font-size: 20px;
+    margin: 0;
+`
+export const ProfileCount = styled.div``
+export const ProfileGrade = styled.div`
+    width: 20%;
+`
+
+export const GradeAmount = styled.h2`
+width: 100%;
+font-size: 15px;
+text-align: center;
+margin: 0;
+`;
+export const Grade = styled.h2`
+width: 100%;
+font-size: 15px;
+text-align: center;
+margin: 0;
+`;
+
+export const ContentBox = styled.div`
+width: 100%;
+margin-bottom: 60px;
+`
+export const ContentSubBox = styled.div`
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 20px;
+`
+
+export const AskedState = styled.div`
+    color: blue;
+    border: 1px solid black;
+    padding: 10px;
+    border-radius: 10px;  
+`
+export const Date = styled.div`
+    color: gray;
+    font-size: 10px;
+`
+export const ContentTitle = styled.h2`
+    margin: 0;
+    margin-bottom: 10px;
+`
+export const ContentHashtag = styled.div`
+    font-size: 13px;
+    color: gray;
+    margin-bottom: 10px;
+`
+export const ContentDescription = styled.div`
+    width: 80%;
+    margin-bottom: 10px;
+`
+export const ContentViewCount = styled.div`
+font-size: 13px;
+color: gray;
+`
+
+export const MoreBox = styled.div`
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+`
+export const PaymentCondition = styled.div`
+    font-size: 13px;
+    color: gray;
+    flex-grow: 1;
+`
+export const ChattingButton = styled.button`
+    width: 30%;
+    height: 40px;
+    color: white;
+    border: none;
+    background-color: rgb(0, 137, 181);
+    border-radius: 10px;
+`
