+++ src/common/fonts/NanumSquareRoundL.ttf
